+++ src/components/Data/Herkunft/index.js
@@ -85,7 +85,7 @@ const herkunftQuery = gql`
   ${herkunftFragment}
 `
 const personFelderQuery = gql`
-  query personFelderQuery($personId: Int) {
+  query personFelderQuery($personId: bigint) {
     person_felder(where: { person_id: { _eq: $personId } }) {
       ...PersonFelderFields
     }
