+++ src/components/Data/Kultur/Settings.js
@@ -46,7 +46,7 @@ const SettingsKultur = ({ kulturResult }) => {
         await client.mutate({
           mutation: gql`
               mutation update_kultur_felder(
-                $kulturId: Int!
+                $kulturId: bigint!
               ) {
                 update_kultur_felder(
                   where: { kultur_id: { _eq: $kulturId } }
